+++ src/nu_combine_methods.h
@@ -25,19 +25,19 @@
 #include "nu_combine_defs.h"
 
 int NU_mean_function(double *data, double *weights,
-    int size, double *out[NU_COMBINE_OUTDIM], void *func_data);
+    size_t size, double *out[NU_COMBINE_OUTDIM], void *func_data);
 
 int NU_median_function(double *data, double *weights,
-    int size, double *out[NU_COMBINE_OUTDIM], void *func_data);
+    size_t size, double *out[NU_COMBINE_OUTDIM], void *func_data);
 
 int NU_minmax_function(double *data, double *weights,
-    int size, double *out[NU_COMBINE_OUTDIM], void *func_data);
+    size_t size, double *out[NU_COMBINE_OUTDIM], void *func_data);
 
 int NU_sigmaclip_function(double *data, double *weights,
-    int size, double *out[NU_COMBINE_OUTDIM], void *func_data);
+    size_t size, double *out[NU_COMBINE_OUTDIM], void *func_data);
 
 int NU_quantileclip_function(double *data, double *weights,
-    int size, double *out[NU_COMBINE_OUTDIM], void *func_data);
+    size_t size, double *out[NU_COMBINE_OUTDIM], void *func_data);
 
 void NU_destructor_function(void* cobject, void *cdata);
